+++ CHANGELOG.rst
@@ -11,6 +11,7 @@ Unreleased
 ------------------
 Fixed
  - Fixed `tcod.map_copy` could break the `tcod.map.Map` class.
+ - Fixed `tcod.map_clear` `transparent` and `walkable` parameters were reverse.
 
 4.5.0 - 2018-06-12
 ------------------